Attempting to improve marshalling of S3 PutObjectRequest.
Most of the AWS SDK Request objects directly extend
com.amazonaws.AmazonWebServiceRequest which has attributes
that are not useful to conversion to and from Clojure data.
So the amazonica.core/accessors function uses Class.getDeclaredMethods()
to reflect on a Class' "bean" attributes without pulling in
many unhelpful fields from the parent class.

In the case of com.amazonaws.services.s3.model.PutObjectRequest,
there is an intermediate parent before the generic
AmazonWebServicesRequest class where most of the useful request
information is located.  Amazonica's default marshalling behavior
is only able to see a single field for this Request class.

This attempts to improve marshalling behavior strictly for mock
handler functions and only when an intermediate class exists
between the actual class and the base abstract AmazonWebServiceRequest
class.